+++ CHANGES.md
@@ -6,6 +6,12 @@ Version 0.1.1
 
 To be released.
 
+ -  Fixed a bug where `object()` parsers containing only Boolean flags would
+    fail when no arguments were provided, instead of defaulting the flags to
+    `false`. [[#6]]
+
+[#6]: https://github.com/dahlia/optique/issues/6
+
 
 Version 0.1.0
 -------------